+++ src/main/factories/service_factory.ts
@@ -1,10 +1,11 @@
-import type { AuthenticationService } from "../../domain/services/authentication_service";
-import type { ContributorService } from "../../domain/services/contributor_service";
-import type { DashboardService } from "../../domain/services/dashboard_service";
+import type { ComplianceRepository } from "../../domain/repositories/compliance_repository";
 import type { ContributorRepository } from "../../domain/repositories/contributor_repository";
 import type { RepositoryRepository } from "../../domain/repositories/repository_repository";
 import type { SonarRepository } from "../../domain/repositories/sonar_repository";
 import type { WakaTimeRepository } from "../../domain/repositories/wakatime_repository";
+import type { AuthenticationService } from "../../domain/services/authentication_service";
+import type { ContributorService } from "../../domain/services/contributor_service";
+import type { DashboardService } from "../../domain/services/dashboard_service";
 import { EncryptedAuthenticationService } from "../../infrastructure/services/encrypted_authentication_service";
 import { LocalStorageAuthenticationService } from "../../infrastructure/services/local_storage_authentication_service";
 import { GitHubDashboardService } from "../../service/github_dashboard_service";
@@ -18,7 +19,8 @@ export const createAuthenticationService = async (): Promise<AuthenticationServi
 export const createDashboardService = (
   repositoryRepository: RepositoryRepository,
   sonarRepository: SonarRepository,
-): DashboardService => new GitHubDashboardService(repositoryRepository, sonarRepository);
+  complianceRepository: ComplianceRepository,
+): DashboardService => new GitHubDashboardService(repositoryRepository, sonarRepository, complianceRepository);
 
 export const createContributorService = (
   contributorRepository: ContributorRepository,