+++ src/service/mappers/graphql_repository_mapper.ts
@@ -63,4 +63,5 @@ export const mapGraphQLNodeToRepository = (node: GraphQLRepositoryNode): Reposit
   hasWorkflows: node.defaultBranchRef?.target.statusCheckRollup != null,
   branches: node.branchRefs.nodes.map((ref) => ref.name),
   sonarMetrics: null,
+  complianceStatus: null,
 });